Add new param apply_volume_limit for deriveNMarkets()
When deriving price across several markets, only the volume of the last
step was used. This may lead to giving more weight to conversion ways
with small volume or liquidity on intermediate steps. A new options
enables to pick the volume of the conversion step with smallest volume
across all steps.